feat: set id and description for VIP-314: upgrade XVSVault
