+++ multisig/simulations/opbnbmainnet/index.ts
@@ -6,6 +6,7 @@ import "./vip-004";
 import "./vip-005";
 import "./vip-008";
 import "./vip-009";
+import "./vip-010";
 import "./vip-011";
 import "./vip-012";
 import "./vip-018";
