+++ multisig/simulations/opbnbtestnet/index.ts
@@ -5,5 +5,6 @@ import "./vip-003";
 import "./vip-004";
 import "./vip-005";
 import "./vip-009";
+import "./vip-010";
 import "./vip-012";
 import "./vip-013";
