+++ multisig/simulations/opbnbtestnet/vip-010/index.ts
@@ -6,31 +6,31 @@ import { NETWORK_ADDRESSES } from "../../../../src/networkAddresses";
 import { calculateMappingStorageSlot } from "../../../../src/utils";
 import { forking, pretendExecutingVip } from "../../../../src/vip-framework";
 import { checkXVSVault } from "../../../../src/vip-framework/checks/checkXVSVault";
-import vip030, {
+import vip010, {
   ACM,
-  ETHEREUM_BLOCKS_PER_YEAR,
+  BLOCKS_PER_YEAR,
   NEW_XVS_IMPLEMENTATION,
   XVS_VAULT_PROXY,
-} from "../../../proposals/ethereum/vip-030";
+} from "../../../proposals/opbnbtestnet/vip-010";
 import XVS_VAULT_ABI from "./abi/XVSVault.json";
 import ACM_ABI from "./abi/accessControlManager.json";
 
-const { ethereum } = NETWORK_ADDRESSES;
+const { opbnbtestnet } = NETWORK_ADDRESSES;
 
-const XVS_ADDRESS = "0xd3CC9d8f3689B83c91b7B59cAB4946B063EB894A";
-const POOL_ID = 0;
+const XVS_ADDRESS = "0xc2931B1fEa69b6D6dA65a50363A8D75d285e4da9";
+const POOL_ID = 1;
 const MAPPING_STORAGE_SLOT = 18;
 
-forking(19923613, async () => {
+// NOTE: cannot find any pending rewards for XVS on this chain neither with PoolID = 0 or with PoolID = 1
+forking(29407404, async () => {
   const provider = ethers.provider;
   let xvsVaultProxy: Contract;
-  let pendingWithdrawalsBefore: BigNumber;
   let accessControlManager: Contract;
+  let pendingWithdrawalsBefore: BigNumber;
 
   before(async () => {
     xvsVaultProxy = new ethers.Contract(XVS_VAULT_PROXY, XVS_VAULT_ABI, provider);
     accessControlManager = await ethers.getContractAt(ACM_ABI, ACM);
-
     const storageSlot = calculateMappingStorageSlot(XVS_ADDRESS, POOL_ID, MAPPING_STORAGE_SLOT);
     const value = await provider.getStorageAt(XVS_VAULT_PROXY, storageSlot);
     pendingWithdrawalsBefore = BigNumber.from(ethers.constants.HashZero === value ? 0 : utils.stripZeros(value));
@@ -38,7 +38,7 @@ forking(19923613, async () => {
 
   describe("Post-VIP behavior", async () => {
     before(async () => {
-      await pretendExecutingVip(vip030());
+      await pretendExecutingVip(vip010());
     });
 
     checkXVSVault();
@@ -49,13 +49,13 @@ forking(19923613, async () => {
 
     it("Xvs vault should be block based with correct number of blocks", async () => {
       expect(await xvsVaultProxy.isTimeBased()).to.be.equal(false);
-      expect(await xvsVaultProxy.blocksOrSecondsPerYear()).to.be.equal(ETHEREUM_BLOCKS_PER_YEAR);
+      expect(await xvsVaultProxy.blocksOrSecondsPerYear()).to.be.equal(BLOCKS_PER_YEAR);
     });
 
     it("Check permission for setRewardAmountPerBlock", async () => {
       expect(
         await accessControlManager.hasPermission(
-          ethereum.GUARDIAN,
+          opbnbtestnet.GUARDIAN,
           XVS_VAULT_PROXY,
           "setRewardAmountPerBlock(address,uint256)",
         ),
@@ -65,7 +65,7 @@ forking(19923613, async () => {
     it("Check permission for setRewardAmountPerBlockOrSecond", async () => {
       expect(
         await accessControlManager.hasPermission(
-          ethereum.GUARDIAN,
+          opbnbtestnet.GUARDIAN,
           XVS_VAULT_PROXY,
           "setRewardAmountPerBlockOrSecond(address,uint256)",
         ),
