+++ multisig/simulations/sepolia/index.ts
@@ -18,3 +18,4 @@ import "./vip-023";
 import "./vip-024";
 import "./vip-026";
 import "./vip-029";
+import "./vip-033";
