--- vips/vip-308/bscmainnet.ts
@@ -1,78 +1,0 @@
-import { ProposalType } from "../../src/types";
-import { makeProposal } from "../../src/utils";
-
-export const NORMAL_TIMELOCK = "0x939bD8d64c0A9583A7Dcea9933f7b21697ab6396";
-export const FAST_TRACK_TIMELOCK = "0x555ba73dB1b006F3f2C7dB7126d6e4343aDBce02";
-export const CRITICAL_TIMELOCK = "0x213c446ec11e45b15a6E29C1C1b402B8897f606d";
-
-export const ACM = "0x4788629ABc6cFCA10F9f969efdEAa1cF70c23555";
-
-export const XVS_VAULT_PROXY = "0x051100480289e704d20e9DB4804837068f3f9204";
-export const NEW_XVS_IMPLEMENTATION = "0x413c1E1b77190BC84717F8cCe6EeAb0594E0af4e";
-
-export const BNB_BLOCKS_PER_YEAR = 10_512_000; // assuming a block is mined every 3 seconds
-
-const vip308 = () => {
-  const meta = {
-    version: "v2",
-    title: "VIP-308 Update XVSVault Implementation",
-    description: ``,
-    forDescription: "Execute this proposal",
-    againstDescription: "Do not execute this proposal",
-    abstainDescription: "Indifferent to execution",
-  };
-
-  return makeProposal(
-    [
-      {
-        target: XVS_VAULT_PROXY,
-        signature: "_setPendingImplementation(address)",
-        params: [NEW_XVS_IMPLEMENTATION],
-      },
-      {
-        target: NEW_XVS_IMPLEMENTATION,
-        signature: "_become(address)",
-        params: [XVS_VAULT_PROXY],
-      },
-      {
-        target: XVS_VAULT_PROXY,
-        signature: "initializeTimeManager(bool,uint256)",
-        params: [false, BNB_BLOCKS_PER_YEAR],
-      },
-      {
-        target: ACM,
-        signature: "revokeCallPermission(address,string,address)",
-        params: [XVS_VAULT_PROXY, "setRewardAmountPerBlock(address,uint256)", NORMAL_TIMELOCK],
-      },
-      {
-        target: ACM,
-        signature: "revokeCallPermission(address,string,address)",
-        params: [XVS_VAULT_PROXY, "setRewardAmountPerBlock(address,uint256)", FAST_TRACK_TIMELOCK],
-      },
-      {
-        target: ACM,
-        signature: "revokeCallPermission(address,string,address)",
-        params: [XVS_VAULT_PROXY, "setRewardAmountPerBlock(address,uint256)", CRITICAL_TIMELOCK],
-      },
-      {
-        target: ACM,
-        signature: "giveCallPermission(address,string,address)",
-        params: [XVS_VAULT_PROXY, "setRewardAmountPerBlockOrSecond(address,uint256)", NORMAL_TIMELOCK],
-      },
-      {
-        target: ACM,
-        signature: "giveCallPermission(address,string,address)",
-        params: [XVS_VAULT_PROXY, "setRewardAmountPerBlockOrSecond(address,uint256)", FAST_TRACK_TIMELOCK],
-      },
-      {
-        target: ACM,
-        signature: "giveCallPermission(address,string,address)",
-        params: [XVS_VAULT_PROXY, "setRewardAmountPerBlockOrSecond(address,uint256)", CRITICAL_TIMELOCK],
-      },
-    ],
-    meta,
-    ProposalType.REGULAR,
-  );
-};
-
-export default vip308;
